academicInsights/extraction_agent/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
python-dotenv==1.0.0
httpx==0.25.1
requests==2.31.0
pydantic>=1.10.0